Remove buttons
1) added remove buttons for each form row in formsets
2) Corrected results backend logic
