+++ resume_app/NoResume/templates/results.html
@@ -26,7 +26,7 @@
             <div id="form_row">{{ form.as_p }}</div>
           {% endfor %}
         </div>
-        <button type="submit">Confirm</button>
+        <button type="submit">Generate</button>
       </form>
     </div>
   </body>
